Minor refactoring and name changes
