+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/resourcecardpage/view/ResourcesView.kt
@@ -1,14 +1,11 @@
 package org.wycliffeassociates.otter.jvm.resourcecardpage.view
 
-import org.wycliffeassociates.otter.jvm.resourcecardpage.viewmodel.ResourcesViewModel
 import tornadofx.View
 import tornadofx.Workspace
 import tornadofx.removeFromParent
 import tornadofx.vbox
 
 class ResourcesView : View() {
-
-    private val viewModel: ResourcesViewModel by inject()
 
     override val root = vbox {}
 
@@ -17,7 +14,7 @@ class ResourcesView : View() {
     init {
         activeFragment.header.removeFromParent()
 
-        activeFragment.dock<ResourcePageFragment>()
+        activeFragment.dock<ResourceListFragment>()
         add(activeFragment)
     }
 }
